% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_confidentialite.R
\name{create_GS_CPP}
\alias{create_GS_CPP}
\title{Create confidential grids}
\usage{
create_GS_CPP(tab, seuil, mailles, agreg = FALSE, ...)
}
\arguments{
\item{tab}{A data.table at individual level with at least (x, y, crs) columns
or an aggregated data.table with at least (n_obs, x, y, crs). In this latter case,
observations can be aggregated on the centroid of the squares of the finest grid level.}

\item{seuil}{The confidentiality threshold.}

\item{mailles}{A vector with the sizes of the different
grids.}

\item{agreg}{Boolean to mention if tab is already an agreggated table.}

\item{...}{other arguments from \code{create_grid_niv()}}
}
\value{
At the end of the process, we get a data.table
with one row for each mesh of each grid. In the columns
we have the identifier of the mesh, the identifier of
the mesh at the above level, the number of observation
in the mesh, the state of the mesh (can the information
on the mesh be disseminated without perturbation or not),
the force of the mesh and the group of the mesh.
}
\description{
The confidentiality rule is not to have
meshs containing less observations than
the threshold.
}
\examples{
library(data.table)
n <- 1e4
tab <- as.data.table(data.frame(id_obs = 1:n, x = rnorm(n, 3e6, 2e4), y = rnorm(n, 2e6, 3e4), crs = 3035))
tab_grid <- create_GS_CPP(tab, 5, c(32e3,16e3,8e3,4e3,2e3,1e3))

}
