% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_confidentialite.R
\name{determiner_car_naturel}
\alias{determiner_car_naturel}
\title{determine the natural grid}
\usage{
determiner_car_naturel(resul_GS)
}
\arguments{
\item{resul_GS}{result of the create_GS_CPP() function}
}
\value{
a data.table dataframe
}
\description{
determine the natural grid
}
\examples{
library(data.table)
n <- 1e4
tab <- as.data.table(
  data.frame(
    id_obs = 1:n,
    x = rnorm(n, 3e6, 2e4),
    y = rnorm(n, 2e6, 3e4),
    crs = 3035)
)
tab_GS <- create_GS_CPP(tab, 5, c(32e3,16e3,8e3,4e3,2e3,1e3))
tab_car_nat <- determiner_car_naturel(tab_GS)

}
