% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_visualisation.R
\name{make_contour_car}
\alias{make_contour_car}
\title{Creates a polygon from tile's ID}
\usage{
make_contour_car(id_car)
}
\arguments{
\item{id_car}{id of tile, in INSPIRE style}
}
\value{
sf object
}
\description{
Creates a polygon from tile's ID
}
\examples{
make_contour_car("CRS3035RES200mN2009400E2992400")

}
