% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_estimation.R
\name{imputer_cle_repartition}
\alias{imputer_cle_repartition}
\title{Estimate protected mesh}
\usage{
imputer_cle_repartition(tab_car, list_var_imput, var_cle)
}
\arguments{
\item{tab_car}{A data.table of the variables agregates in a grid.}

\item{var_cle}{A character, the name of the non-sensitive variable
according to which the distribution is done.}

\item{list_var}{A character vector giving the names of the variables
to estimate.}
}
\value{
The data.table \code{tab_car} with the estimated variables.
}
\description{
The information on some mesh cannot
be disseminated because it doesn't respect the threshold rule.
For those mesh, an estimation is done by distributing the total
of the variable on a bigger zone (the groupe) into the meshes,
proportionnaly to a given variable (non sensitive variable).
}
